requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
django==2.1
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
mysqlclient==1.3.13
pycparser==2.19
pymysql==0.9.2
pytz==2018.7
requests==2.19.1
six==1.12.0
urllib3==1.23
ipython==5.1.0